you hdd or ssd passwort
i have experience with password wiping a locked hdd and also ssd. i can protect your ssd hard disk or normal hard disk for companies that if a hacker enters a biso password for the hard disk via the internet, this locks the hard disk. that with my tool the company can open the hdd again,